Added reading regarding AR Associations to Resources section in README.md
I've added a helpful reading a new "Resources" section at the bottom of README.md. This link was very helpful for myself and a few of my classmates and I think it would fit well here. Please do move it to a different lesson/lab if you think it would fit better somewhere else.